refactor(check-in): reorganizes data layer
