+++ src/components/KeyResult/CheckInForm/Fields/CurrentProgress/current-progress.tsx
@@ -6,7 +6,7 @@ import { useRecoilValue } from 'recoil'
 
 import { selectMaskBasedOnFormat } from 'src/components/KeyResult/NumberMasks/selectors'
 import { KeyResult } from 'src/components/KeyResult/types'
-import { buildPartialSelector } from 'src/state/recoil/key-result/selectors'
+import buildPartialSelector from 'src/state/recoil/key-result/build-partial-selector'
 
 import { CheckInFormValues } from '../../form'
