+++ src/components/KeyResult/ProgressSlider/slider.tsx
@@ -6,15 +6,13 @@ import SliderWithHoverThumb, {
 } from 'src/components/Base/SliderWithHoverThumb'
 import { KeyResult } from 'src/components/KeyResult/types'
 import useConfidenceTag from 'src/state/hooks/useConfidenceTag'
+import buildPartialSelector from 'src/state/recoil/key-result/build-partial-selector'
 import {
   keyResultCheckInPopoverOpen,
   keyResultCheckInProgressDraft,
   keyResultCheckInProgressSliderStep,
 } from 'src/state/recoil/key-result/check-in'
-import {
-  buildPartialSelector,
-  selectCurrentConfidence,
-} from 'src/state/recoil/key-result/selectors'
+import selectCurrentConfidence from 'src/state/recoil/key-result/current-confidence'
 
 export interface ProgressSliderSliderProperties {
   keyResultID?: KeyResult['id']
