+++ src/state/recoil/key-result/build-partial-selector.spec.ts
