+++ src/state/recoil/key-result/check-in/progress-draft.ts
@@ -1,7 +1,7 @@
 import { atomFamily } from 'recoil'
 
 import { KeyResult } from 'src/components/KeyResult/types'
-import selectCurrentProgress from 'src/state/recoil/key-result/selectors/current-progress'
+import selectCurrentProgress from 'src/state/recoil/key-result/current-progress'
 
 import { PREFIX } from './constants'
