+++ src/state/recoil/key-result/index.ts
@@ -9,3 +9,11 @@ export * as keyResultDrawer from './drawer'
 export * as keyResultTimeline from './timeline'
 
 export { default as keyResultAtomFamily } from './atom-family'
+
+export { default as keyResultSelector } from './key-result'
+
+export { default as keyResultCurrentProgress } from './current-progress'
+
+export { default as keyResultCurrentConfidence } from './current-confidence'
+
+export { default as keyResultBuildPartialSelector } from './build-partial-selector'
