--- src/state/recoil/key-result/selectors/constants.ts
@@ -1,3 +1,0 @@
-import { PREFIX as PARENT_PREFIX } from '../constants'
-
-export const PREFIX = `${PARENT_PREFIX}::SELECTORS`
