--- src/state/recoil/key-result/selectors/index.ts
@@ -1,13 +1,0 @@
-export { default as buildPartialSelector } from './build-partial-selector'
-
-export { default as selectKeyResult } from './key-result'
-
-export { default as selectCurrentProgress } from './current-progress'
-
-export { default as selectCurrentConfidence } from './current-confidence'
-
-export { default as selectLatestCheckIn } from './latest-check-in'
-
-export { default as selectLatestTimelineEntry } from './latest-timeline-entry'
-
-export { default as removeTimelineEntry } from './remove-timeline-entry'
