+++ src/state/recoil/key-result/timeline/index.ts
@@ -1,1 +1,5 @@
 export { default as keyResultTimelineFetched } from './fetched'
+
+export { default as keyResultTimelineLatestEntry } from './latest-entry'
+
+export { default as keyResultTimelineRemoveEntry } from './remove-entry'
